overlay: maintain client requested size and aligned buffer size
When a client makes a request for memory to overlay we need to
keep track of the size distinctly. This is required for secure
playback use cases in which we internally calculate the aligned
buffer size. The decision to reallocate memory should always be
based on the original size requested by the client and not our
internally calculated aligned buffer size.

Change-Id: I731560be7dba27264d00fdcb1a12622d604ce0fb