hwc: Fix framebuffer resolution for virtual display
Rely on content list to deduce width and height of frame buffer target.
In case of WFD/Screen Record/Screen cast, output buffer and framebuffer
are configured to different formats. In such cases, Panel resolution
reflects output buffer configuration and might not be same as that of
framebuffer target.

CRs-Fixed: 880367
Change-Id: I58b09a1053ce7625c5f7b08f1f943d6b16040ed2