hwc: Handle SBM operation set for all layers
SBM (single buffer mode) operation can be set by graphics on all
layers including FB with a default value of 0. So far HWC assumed
that when an operation followed by 0 value is seen, the feature
should be turned off. The new behavior is to allow multiple
SBM sets with 0 value.

For example: On a genuine SBM session, HWC sees the operation flag
set and a value 1, thus keeps the feature on. When a volume bar
comes up, it has the SBM operation flag as well, but value is 0.
HWC should take care of this and not turn off the feature.

CRs-fixed: 937179
Change-Id: Id3b787cd92636a8ac9be4521b0b5c38327f60bb5