hwc2: Always set the retire fence in present()
When an empty layer set is received in HWC, the flush_ flag is
set. In this case the retire fence in that present call in not
set even though the present ends successfully.
Also, in the flush_ case, set the stored retire fence to -1 as we
don't get one from the driver.

Change-Id: I7953878e7f5c95557c6cf7d70f7b41bc1696d1da
CRs-Fixed: 1030993