sdm: Add option to defer video mode retire fence
Per the HWC2 header the present(retire) fence should be signaled at
the vsync when the result of composition of *this* frame starts to
appear (for video-mode panels). In our current implementation,
the retire fence timeline is the same as the release fence
timeline for video mode, so the retire fence that we get at present()
signals when the next frame comes in to replace the current one.
Hence, for video mode panels, we need to defer the retire fence
to meet the header requirements.

Change-Id: Ia84c9f14ba5c3667c4cad996027638f1f074914f
CRs-Fixed: 1050596