sdm: add check for disabling external animation
In current design the flag passed from clients is used to
disable external animation. Refine this flag to two parts,
one is to indicate whether animation is present, another
is to check the system property and disable external
animation. So other functions can know the presence of
animation layers if necessary.

CRs-fixed: 1019601

Change-Id: Ib6ffb7a28d6935b86bc93abc82049b77b3ddcd08