hwc: Fix locking on event handling and validation state
Fix a small gap in locking on event handling where validation
state was reset after lock scope was over. Also validation state
was being reset for all displays even if event was for primary.
This is incorrect since lock was only held for primary. Move
validation state reset to be per display

Change-Id: I24a8930d536a906d2b8adba35833774c03b59933
CRs-fixed: 1114808