Use 'subdir-objects' option in HAL configure options
If a source file is in a subdirectory, but the 'subdir-objects'
automake option hasn't been enabled. For now, the corresponding
output object file(s) will be placed in the top-level directory.
However, this behaviour will change in future Automake versions,
they will unconditionally cause object files to be placed in the
same subdirectory of the corresponding sources.
So to avoid future incompatibilities, enable the 'subdir-objects'
option in location module configure options.

Change-Id: If9e94c0f6fcd9922ee7f0f9d1909d5d6579a5715
CRs-Fixed: 2533108